pipeline/requirements.txt
groq
pdfplumber
python-docx
python-dotenv
tenacity
